+++ pyproject.toml
@@ -28,6 +28,7 @@ argon2-cffi = "^23.1.0"
 typer = "^0.12.5"
 prometheus-client = "^0.21.0"
 openrelik-ai-common = "~0.5.0"
+opentelemetry-instrumentation-celery = "^0.53b1"
 opentelemetry-instrumentation-fastapi = "^0.53b1"
 opentelemetry-exporter-otlp = "1.32.1"
