+++ src/api/v1/taskqueue.py
@@ -17,10 +17,14 @@ import os
 from celery.app import Celery
 from fastapi import APIRouter
 
+from opentelemetry.instrumentation.celery import CeleryInstrumentor
 from lib import celery_utils
+from lib import tracing
 
 redis_url = os.getenv("REDIS_URL")
 celery = Celery(broker=redis_url, backend=redis_url)
+tracing.setup_telemetry(service_name='openrelik-server-task-queue')
+CeleryInstrumentor().instrument(celery_app=celery) # <-- THE CRUCIAL FIX
 
 router = APIRouter()
