+++ src/main.py
@@ -47,17 +47,13 @@ from datastores.sql.models.user import User
 from healthz import router as healthz_router
 from lib import celery_utils
 
+from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
+from opentelemetry.instrumentation.celery import CeleryInstrumentor
+
 telemetry_enabled = os.environ.get("OTEL_MODE", '') != ''
 
 if telemetry_enabled:
-    from opentelemetry import trace
-
-    from opentelemetry.exporter.otlp.proto.grpc import trace_exporter as grpc_exporter
-    from opentelemetry.exporter.otlp.proto.http import trace_exporter as http_exporter
-    from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
-    from opentelemetry.sdk.resources import Resource
-    from opentelemetry.sdk.trace import TracerProvider
-    from opentelemetry.sdk.trace.export import BatchSpanProcessor
+    from lib import tracing
 
 # Allow Frontend origin to make API calls.
 origins = config["server"]["allowed_origins"]
@@ -100,40 +96,11 @@ async def lifespan(app: FastAPI):
     pass
 
 
-def setup_telemetry(service_name: str):
-    """Configures the OpenTelemetry trace exporter.
-
-    Args:
-        service_name (str): the service name used to identify generated traces.
-    """
-
-    resource = Resource(attributes={
-        "service.name": service_name
-    })
-
-    # --- Tracing Setup ---
-    trace.set_tracer_provider(TracerProvider(resource=resource))
-
-    otel_mode = os.environ.get("OTEL_MODE", "otlp-grpc")
-    otlp_grpc_endpoint = os.environ.get("OTLP_GRPC_ENDPOINT", "jaeger-collector:4317")
-    otlp_http_endpoint = os.environ.get("OTLP_HTTP_ENDPOINT", "http://jaeger-collector:4318/v1/traces")
-
-    trace_exporter = None
-    if otel_mode == "otlp-grpc":
-        trace_exporter = grpc_exporter.OTLPSpanExporter(
-                endpoint=otlp_grpc_endpoint, insecure=True)
-    elif otel_mode == "otlp-http":
-        trace_exporter = http_exporter.OTLPSpanExporter(
-                endpoint=otlp_http_endpoint)
-    else:
-        raise Exception("Unsupported OTEL tracing mode %s", otel_mode)
-
-    trace.get_tracer_provider().add_span_processor(BatchSpanProcessor(trace_exporter))
 
 
 
 if telemetry_enabled:
-    setup_telemetry("openrelik-server")
+    tracing.setup_telemetry("openrelik-server")
 
 # Create the main app
 app = FastAPI(lifespan=lifespan)
@@ -259,11 +226,13 @@ api_v1.include_router(
     ],
 )
 
-if telemetry_enabled:
-    FastAPIInstrumentor.instrument_app(api_v1)
-
 # Setup the queues. This function take all registered tasks on the celery task queue
 # and generate the task queue config automatically.
 redis_url = os.getenv("REDIS_URL")
 celery = Celery(broker=redis_url, backend=redis_url)
+tracing.setup_telemetry(service_name='openrelik-server')
+FastAPIInstrumentor.instrument_app(api_v1)
+CeleryInstrumentor().instrument(celery_app=celery)
+
 celery_utils.update_task_queues(celery)
+
